+++ lib/thrift_server.rb
@@ -14,6 +14,10 @@ require_relative 'thrift_server/server_metrics_subscriber'
 require_relative 'thrift_server/rpc_metrics_subscriber'
 require_relative 'thrift_server/log_subscriber'
 
+require_relative 'thrift_server/publisher'
+
+require_relative 'thrift_server/server'
+
 module ThriftServer
   RPC = Struct.new(:name, :args, :exceptions) do
     def initialize(*)
@@ -67,111 +71,6 @@ module ThriftServer
     def finalize_stack!
       stack.use Dispatcher, handler
       stack.finalize!
-    end
-  end
-
-  class Publisher
-    include Enumerable
-
-    extend Forwardable
-
-    def_delegators :listeners, :each
-
-    attr_reader :listeners
-
-    def initialize
-      @listeners = [ ]
-    end
-
-    def subscribe(object)
-      listeners << object
-    end
-
-    def publish(event, *args)
-      listeners.each do |listener|
-        listener.send(event, *args) if listener.respond_to? event
-      end
-    end
-  end
-
-  class Server < Thrift::ThreadPoolServer
-    extend Forwardable
-
-    def_delegators :@processor, :use
-    def_delegators :@processor, :publisher, :publish, :subscribe
-
-    attr_accessor :port
-
-    def log(logger)
-      subscribe LogSubscriber.new(logger)
-    end
-
-    def metrics(statsd)
-      subscribe ServerMetricsSubscriber.new(statsd)
-      subscribe RpcMetricsSubscriber.new(statsd)
-    end
-
-    def threads
-      @thread_q.max
-    end
-
-    def protocol
-      @protocol_factory
-    end
-
-    def transport
-      @transport_factory
-    end
-
-    def server_transport
-      @server_transport
-    end
-
-    def start(dry_run: false)
-      publish :server_start, self
-
-      serve unless dry_run
-    end
-
-    def serve
-      @server_transport.listen
-
-      begin
-        loop do
-          @thread_q.push(:token)
-          publish :server_thread_pool_change, delta: 1
-
-          Thread.new do
-            begin
-              loop do
-                client = @server_transport.accept
-                remote_address = client.handle.remote_address
-
-                publish :server_connection_opened, remote_address
-
-                trans = @transport_factory.get_transport(client)
-                prot = @protocol_factory.get_protocol(trans)
-                begin
-                  loop do
-                    @processor.process(prot, prot)
-                  end
-                rescue Thrift::TransportException, Thrift::ProtocolException => e
-                  publish :server_connection_closed, remote_address
-                ensure
-                  trans.close
-                end
-              end
-            rescue => e
-              @exception_q.push(e)
-            ensure
-              publish :server_thread_pool_change, delta: -1
-              @thread_q.pop # thread died!
-            end
-          end
-        end
-      ensure
-        @server_transport.close
-      end
     end
   end
